refactored implementation, tested functionality.
